refactor: utils 폴더 restaurantUtils로 이름 변경
